move src code to subfolder
