+++ .flake8
@@ -16,10 +16,10 @@ exclude =
   venv*/
   target
   __pycache__
-  exporter/opentelemetry-exporter-jaeger-proto/src/opentelemetry/exporter/jaeger/gen/
+  exporter/opentelemetry-exporter-jaeger-proto/src/opentelemetry/exporter/jaeger/proto/gen/
   exporter/opentelemetry-exporter-jaeger-proto/build/*
-  exporter/opentelemetry-exporter-jaeger-proto/src/opentelemetry/exporter/jaeger/gen/
-  exporter/opentelemetry-exporter-jaeger-proto/build/*
+  exporter/opentelemetry-exporter-jaeger-thrift/src/opentelemetry/exporter/jaeger/thrift/gen/
+  exporter/opentelemetry-exporter-jaeger-thrift/build/*
   exporter/opentelemetry-exporter-zipkin/src/opentelemetry/exporter/zipkin/encoder/v2/protobuf/gen/
   docs/examples/opentelemetry-example-app/src/opentelemetry_example_app/grpc/gen/
   docs/examples/opentelemetry-example-app/build/*
