+++ exporter/opentelemetry-exporter-jaeger-proto/src/opentelemetry/exporter/jaeger/proto/translate/protobuf.py
@@ -17,8 +17,8 @@ from typing import Optional, Sequence
 from google.protobuf.duration_pb2 import Duration
 from google.protobuf.timestamp_pb2 import Timestamp
 
-from opentelemetry.exporter.jaeger.gen import model_pb2
-from opentelemetry.exporter.jaeger.translate import (
+from opentelemetry.exporter.jaeger.proto.gen import model_pb2
+from opentelemetry.exporter.jaeger.proto.translate import (
     NAME_KEY,
     OTLP_JAEGER_SPAN_KIND,
     VERSION_KEY,
