+++ exporter/opentelemetry-exporter-jaeger-proto/src/opentelemetry/exporter/jaeger/proto/util.py
