+++ exporter/opentelemetry-exporter-jaeger-proto/src/opentelemetry/exporter/jaeger/proto/version.py
