+++ exporter/opentelemetry-exporter-jaeger-thrift/src/opentelemetry/exporter/jaeger/thrift/gen/zipkincore/ZipkinCollector-remote
