+++ exporter/opentelemetry-exporter-jaeger-thrift/src/opentelemetry/exporter/jaeger/thrift/send/thrift.py
@@ -20,8 +20,8 @@ import socket
 from thrift.protocol import TBinaryProtocol, TCompactProtocol
 from thrift.transport import THttpClient, TTransport
 
-from opentelemetry.exporter.jaeger.gen.agent import Agent as agent
-from opentelemetry.exporter.jaeger.gen.jaeger import Collector as jaeger
+from opentelemetry.exporter.jaeger.thrift.gen.agent import Agent as agent
+from opentelemetry.exporter.jaeger.thrift.gen.jaeger import Collector as jaeger
 
 UDP_PACKET_MAX_LENGTH = 65000
