+++ pyproject.toml
@@ -3,8 +3,8 @@ line-length = 79
 exclude = '''
 (
   /(  # generated files
-    exporter/opentelemetry-exporter-jaeger-proto/src/opentelemetry/exporter/jaeger/gen|
-    exporter/opentelemetry-exporter-jaeger-thrift/src/opentelemetry/exporter/jaeger/gen|
+    exporter/opentelemetry-exporter-jaeger-proto/src/opentelemetry/exporter/jaeger/proto/gen|
+    exporter/opentelemetry-exporter-jaeger-thrift/src/opentelemetry/exporter/jaeger/thrift/gen|
     exporter/opentelemetry-exporter-zipkin/src/opentelemetry/exporter/zipkin/encoder/v2/protobuf/gen|
     opentelemetry-proto/src/opentelemetry/proto/collector|
     opentelemetry-proto/src/opentelemetry/proto/common|
